requirements.txt
streamlit>=1.20.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python-headless>=4.5.5.62
numpy>=1.21.0
Pillow>=9.0.0
